+++ examples/cra-ui/src/index.tsx
@@ -26,6 +26,7 @@ import UiServicePage from './pages/service';
 import { AlertPage } from './pages/alert';
 import { SliderPage } from './pages/slider';
 import { DatetimePage } from './pages/datetime';
+import { ContextMenuPage } from './pages/contextmenu';
 
 const container = document.getElementById('root');
 const root = createRoot(container!); // createRoot(container!) if you use TypeScript
@@ -55,7 +56,7 @@ root.render(
           <Link href="/checkbox">Checkbox</Link> | <Link href="/tag">Tag</Link> | <Link href="/modal">Modal</Link> |{' '}
           <Link href="/wizard">Wizard</Link> | <Link href="/form">Form</Link> |{' '}
           <Link href="/actionmenu">Action Menu</Link> | <Link href="/notifications">Notifications</Link> |{' '}
-          <Link href="/timer">Timer</Link> | <Link href="/service">Service</Link> | <Link href="/alert">Alert</Link>  | <Link href="/datetime">DateTime</Link>
+          <Link href="/timer">Timer</Link> | <Link href="/service">Service</Link> | <Link href="/alert">Alert</Link>  | <Link href="/datetime">DateTime</Link> | <Link href="/contextmenu">ContextMenu</Link>
         </div>
         <div style={{ paddingLeft: '100px', paddingRight: '100px' }}>
           <Routes>
@@ -84,6 +85,7 @@ root.render(
             <Route path="/alert" element={<AlertPage />} />
             <Route path="/slider" element={<SliderPage />} />
             <Route path="/datetime" element={<DatetimePage />} />
+            <Route path="/contextmenu" element={<ContextMenuPage />} />
           </Routes>
         </div>
       </div>