+++ packages/onekijs-ui/src/components/contextmenu/typings.ts
@@ -1,0 +1,6 @@
+export type ContextMenuProps = {
+  className?: string;
+  refElement: HTMLElement | React.RefObject<HTMLDivElement>;
+  offsetX?: number;
+  offsetY?: number;
+};