+++ examples/cra-examples/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.21.9](https://github.com/oneki/onekijs/compare/v0.21.8...v0.21.9) (2026-03-12)
+
+**Note:** Version bump only for package cra-examples
+
+
+
+
+
+## [0.21.8](https://github.com/oneki/onekijs/compare/v0.21.7...v0.21.8) (2026-03-12)
+
+**Note:** Version bump only for package cra-examples
+
+
+
+
+
+## [0.21.7](https://github.com/oneki/onekijs/compare/v0.21.6...v0.21.7) (2026-03-12)
+
+**Note:** Version bump only for package cra-examples
+
+
+
+
+
+## [0.21.6](https://github.com/oneki/onekijs/compare/v0.21.5...v0.21.6) (2026-03-12)
+
+**Note:** Version bump only for package cra-examples
+
+
+
+
+
+## [0.21.5](https://github.com/oneki/onekijs/compare/v0.21.4...v0.21.5) (2026-03-12)
+
+**Note:** Version bump only for package cra-examples
+
+
+
+
+
 ## [0.21.4](https://github.com/oneki/onekijs/compare/v0.21.3...v0.21.4) (2025-10-21)
 
 **Note:** Version bump only for package cra-examples